feat(media): add s3 to manage media files
Users may choose between filesystem (FS) or S3 to store and manage their media files

+++ modules/Media/Models/MediaModel.php
@@ -8,18 +8,18 @@ declare(strict_types=1);
  * @link       https://castopod.org/
  */
 
-namespace App\Models;
+namespace Modules\Media\Models;
 
-use App\Entities\Media\Audio;
-use App\Entities\Media\Chapters;
-use App\Entities\Media\Document;
-use App\Entities\Media\Image;
-use App\Entities\Media\Transcript;
-use App\Entities\Media\Video;
 use CodeIgniter\Database\BaseResult;
 use CodeIgniter\Database\ConnectionInterface;
 use CodeIgniter\Model;
 use CodeIgniter\Validation\ValidationInterface;
+use Modules\Media\Entities\Audio;
+use Modules\Media\Entities\Chapters;
+use Modules\Media\Entities\Document;
+use Modules\Media\Entities\Image;
+use Modules\Media\Entities\Transcript;
+use Modules\Media\Entities\Video;
 
 class MediaModel extends Model
 {
@@ -52,7 +52,7 @@ class MediaModel extends Model
      */
     protected $allowedFields = [
         'id',
-        'file_path',
+        'file_key',
         'file_size',
         'file_mimetype',
         'file_metadata',
@@ -86,26 +86,14 @@ class MediaModel extends Model
         ConnectionInterface &$db = null,
         ValidationInterface $validation = null
     ) {
-        switch ($fileType) {
-            case 'audio':
-                $this->returnType = Audio::class;
-                break;
-            case 'video':
-                $this->returnType = Video::class;
-                break;
-            case 'image':
-                $this->returnType = Image::class;
-                break;
-            case 'transcript':
-                $this->returnType = Transcript::class;
-                break;
-            case 'chapters':
-                $this->returnType = Chapters::class;
-                break;
-            default:
-                // do nothing, keep Document class as default
-                break;
-        }
+        $this->returnType = match ($fileType) {
+            'audio' => Audio::class,
+            'video' => Video::class,
+            'image' => Image::class,
+            'transcript' => Transcript::class,
+            'chapters' => Chapters::class,
+            default => Document::class
+        };
 
         parent::__construct($db, $validation);
     }
@@ -135,8 +123,15 @@ class MediaModel extends Model
      */
     public function saveMedia(object $media): int | false
     {
+        // save file first
+        if (! $media->saveFile()) {
+            return false;
+        }
+
         // insert record in database
         if (! $mediaId = $this->insert($media, true)) {
+            $this->db->transRollback();
+
             return false;
         }
 
@@ -148,6 +143,11 @@ class MediaModel extends Model
      */
     public function updateMedia(object $media): bool
     {
+        // save file first
+        if (! $media->saveFile()) {
+            return false;
+        }
+
         return $this->update($media->id, $media);
     }
 
@@ -166,9 +166,14 @@ class MediaModel extends Model
         return $result;
     }
 
-    public function deleteMedia(object $media): bool|BaseResult
+    /**
+     * @param Document|Audio|Video|Image|Transcript|Chapters $media
+     */
+    public function deleteMedia($media): bool|BaseResult
     {
-        $media->deleteFile();
+        if (! $media->deleteFile()) {
+            return false;
+        }
 
         return $this->delete($media->id);
     }
